Working on fixing virtual vals in object (or virtual methods?)
Camlp5 has `CgMth`, `CgVal` and `CgVir` in class signature item which
correspond to class'es methods, fields and some virtual stuff (?).
In compiler's parsetree there are only fields and methods, and both
can be virtual. So, it is unobvious what is the purpose of the third
constructor.

In this patch dumping of `CgVir` was changed to be a virtual method
instead of virtual val (to be in consistency with pr_o).