off-by-one error in raw string lexing
Previously, for a raw-string literal {x|ab|x} we end up with "ab|x"
in the buffer, but don't add the closing "}".  But when we extract the
string from the buffer, we subtract the length of the entire delimiter
"|x}".  To make this easy, always add chars as we read them, to the
buffer.  Thus the buffer will be "ab|x}" at the moment we've completed
lexing the raw string literal.

Also added a little check to raw_string_starter_p: see the comment
which should be explanatory.