fixed eq_symbol grammar bug.
Bug was causing failure of left-factorization for rules with symbols
that were comprised of rules (e.g. OPT).  Fix is to add the hash of
the action to the rule in pa_extend, then use that in eq_symbol for
equality check.  Could use a little more refactoring, and maybe
switch from hash to the actual AST of the action, but this would
require more reorg of code, so leaving it aside for now.