converted eprinter to support built-in recursion for parens.
Added two new functions defined in each semantic action:

"top": this invokes the top of the current printer.

"bottom": this takes a labeled "fail" argument, which is
a function unit->string, and invoked if an error is detected.

printers now implicitly carry along a "prev" value, which is set
when a semantic action invokes the "bottom" function, to whatever
the AST is at that point.  So if the last rule in a printer invokes
the "bottom" function, then that value is carried in "prev".  The
"curr" and "next" functions carry along "prev", too.  When "bottom"
is invoked on an AST, if it is the same as in "prev", an error is
raised (with the "fail" above); otherwise, prev is set to that AST,
and "top" is invoked.  So effectively, we get a recursion, but that
recursion will terminate after a single loop.

Used this to clean up pr_o.ml