changed revised syntax so attribute-id cna be a quoted string.
Also, when printing in revised syntax, attribute-ides are always
printed as quoted-sstrings.  Parallel change made to q_MLast.  This
was needed so that we can parse attrs like [@value 2].  This works
in original syntax since "value" isn't a keyword, but in revised
syntax, it is a keyword.

We can still write attribute-ids without quotes, but they're always
printed with quotes.