fix "loc" generation in q_ast for generic_constructor
when Q_ast generates patterns, the first arg in every constructor is
"_"; but then a subsequent pass replaces that with "loc" if the
quotation specifies that it wants this (e.g. "<:expr:< ... >>").  But
this doesn't work for generic_constructor, which is a tuple and not a
sum-type.  So fix this.