added just a little bit of fallback pretty-printer support in eprinter
when Eprinter fails, you get no good indication of why. So in
eprinter, make, I added a "fail" function, which can be hooked to the
printers in Pp_debug (which can be fulfilled by pa_ppx.base).  It gets
called when Eprinter fails to print, to format the argument, for
inclusion in the exception.