Use `/usr/bin/env` to invoke bash
The shellbang `#!/bin/bash` is not a portable way to invoke bash. The latest versions of
camlp5 cannot be installed on NixOS because NixOS does not install bash
in `/bin`. `#!/usr/bin/bash env` is not POSIX but it is more portable.

I also replace `SHELL=/bin/bash` by `SHELL=bash`. It is the
responsability of the user to have bash in its PATH.